+++ src/api/org/jzy3d/colors/colormaps/ColorMapRBG.java
@@ -19,13 +19,13 @@ public class ColorMapRBG implements IColorMap {
     }
 
     
-    public Color getColor( IColorMappable colorable, float x, float y, float z ){
+    public Color getColor( IColorMappable colorable, double x, double y, double z ){
         return getColor( x, y, z, colorable.getMin(), colorable.getMax() );        
     }
     
-    private Color getColor( float x, float y, float z, float zMin, float zMax ){
+    private Color getColor( double x, double y, double z, double zMin, double zMax ){
         
-        float rel_value = 0;
+        double rel_value = 0;
         
         if( z < zMin )
             rel_value = 0;
@@ -45,7 +45,7 @@ public class ColorMapRBG implements IColorMap {
         return new Color( r, v, b );
     }
     
-    public Color getColor( IColorMappable colorable, float z ){
+    public Color getColor( IColorMappable colorable, double z ){
         return getColor( 0.0f, 0.0f, z, colorable.getMin(), colorable.getMax() );        //To re-use the existing code
     }
 
@@ -89,7 +89,7 @@ public class ColorMapRBG implements IColorMap {
      * @param bottomwidth
      * @return color influence.
      */
-    private float colorComponentRelative( float value, float center, float topwidth, float bottomwidth ){
+    private double colorComponentRelative( double value, double center, double topwidth, double bottomwidth ){
         return colorComponentAbsolute( value, center-(bottomwidth/2), center+(bottomwidth/2), center-(topwidth/2), center+(topwidth/2) );
     }
     
@@ -116,8 +116,8 @@ public class ColorMapRBG implements IColorMap {
      * 
      * @return color influence.
      */
-    private float colorComponentAbsolute( float value, float bLeft, float bRight, float tLeft, float tRight ){
-        float output = 0;
+    private double colorComponentAbsolute( double value, double bLeft, double bRight, double tLeft, double tRight ){
+        double output = 0;
         // a gauche ou a droite du creneau
         if( (value < bLeft) || (value >= bRight) ){
             output = 0;
