+++ src/api/org/jzy3d/colors/colormaps/ColorMapRainbow.java
@@ -44,7 +44,7 @@ public class ColorMapRainbow implements IColorMap {
     }
 
     
-    public Color getColor(IColorMappable colorable, float x, float y, float z) {
+    public Color getColor(IColorMappable colorable, double x, double y, double z) {
         return getColor( x, y, z, colorable.getMin(), colorable.getMax() );        
     }
     
@@ -52,9 +52,9 @@ public class ColorMapRainbow implements IColorMap {
      * A helper for getColor( ColorMappable waferview, Point3d pt ) that calls
      * other helper functions
      */
-    private Color getColor( float x, float y, float z, float zMin, float zMax ){
+    private Color getColor( double x, double y, double z, double zMin, double zMax ){
         
-        float rel_value = 0;
+        double rel_value = 0;
         
         if( z < zMin )
             rel_value = 0;
@@ -74,7 +74,7 @@ public class ColorMapRainbow implements IColorMap {
         return new Color( r, v, b );
     }
     
-    public Color getColor( IColorMappable colorable, float z ){
+    public Color getColor( IColorMappable colorable, double z ){
         return getColor( 0.0f, 0.0f, z, colorable.getMin(), colorable.getMax() );        //To re-use the existing code
     }
      
@@ -118,7 +118,7 @@ public class ColorMapRainbow implements IColorMap {
      * @param bottomwidth
      * @return color influence.
      */
-    private float colorComponentRelative( float value, float center, float topwidth, float bottomwidth ){
+    private double colorComponentRelative( double value, double center, double topwidth, double bottomwidth ){
         return colorComponentAbsolute( value, center-(bottomwidth/2), center+(bottomwidth/2), center-(topwidth/2), center+(topwidth/2) );
     }
     
@@ -145,8 +145,8 @@ public class ColorMapRainbow implements IColorMap {
      * 
      * @return color influence.
      */
-    private float colorComponentAbsolute( float value, float bLeft, float bRight, float tLeft, float tRight ){
-        float output = 0;
+    private double colorComponentAbsolute( double value, double bLeft, double bRight, double tLeft, double tRight ){
+        double output = 0;
         // a gauche ou a droite du creneau
         if( (value < bLeft) || (value >= bRight) ){
             output = 0;
