+++ src/api/org/jzy3d/colors/colormaps/ColorMapWhiteRed.java
@@ -31,7 +31,7 @@ public class ColorMapWhiteRed implements IColorMap {
     }
 
     
-    public Color getColor( IColorMappable colorable, float x, float y, float z ){
+    public Color getColor( IColorMappable colorable, double x, double y, double z ){
         return getColor( x, y, z, colorable.getMin(), colorable.getMax() );        
     }
     
@@ -39,9 +39,9 @@ public class ColorMapWhiteRed implements IColorMap {
      * A helper for getColor( ColorMappable waferview, Point3d pt ) that calls
      * other helper functions
      */
-    private Color getColor( float x, float y, float z, float zMin, float zMax ){
+    private Color getColor( double x, double y, double z, double zMin, double zMax ){
         
-        float rel_value = 0;
+        double rel_value = 0;
         
         if( z < zMin )
             rel_value = 0;
@@ -54,10 +54,10 @@ public class ColorMapWhiteRed implements IColorMap {
         		rel_value = ( zMax - z ) / ( zMax - zMin );
         }
         
-        return new Color( 1.0f, rel_value, rel_value );
+        return new Color( 1.0f, (float)rel_value, (float)rel_value );
     }
     
-    public Color getColor( IColorMappable colorable, float z ){
+    public Color getColor( IColorMappable colorable, double z ){
         return getColor( 0.0f, 0.0f, z, colorable.getMin(), colorable.getMax() );        //To re-use the existing code
     }
