+++ src/api/org/jzy3d/plot3d/primitives/axes/layout/providers/AbstractTickProvider.java
@@ -3,7 +3,7 @@ package org.jzy3d.plot3d.primitives.axes.layout.providers;
 
 public abstract class AbstractTickProvider implements ITickProvider{
 	@Override
-	public float[] generateTicks(float min, float max) {
+	public double[] generateTicks(double min, double max) {
 		return generateTicks( min, max, getDefaultSteps() );
 	}
 }
