+++ src/api/org/jzy3d/plot3d/primitives/axes/layout/providers/ITickProvider.java
@@ -1,7 +1,7 @@
 package org.jzy3d.plot3d.primitives.axes.layout.providers;
 
 public interface ITickProvider {
-	public float[] generateTicks(float min, float max);
-	public float[] generateTicks(float min, float max, int steps);
+	public double[] generateTicks(double min, double max);
+	public double[] generateTicks(double min, double max, int steps);
 	public int getDefaultSteps();
 }
