+++ src/api/org/jzy3d/plot3d/primitives/axes/layout/providers/SmartTickProvider.java
@@ -15,35 +15,35 @@ public class SmartTickProvider extends AbstractTickProvider implements ITickProv
 	}
 	
 	@Override
-	public float[] generateTicks(float min, float max, int steps) {
+	public double[] generateTicks(double min, double max, int steps) {
 		if(min == max){
-			float[] ticks = new float[1];
+			double[] ticks = new double[1];
 			ticks[0] = min;
 			return ticks;
 		}
 		else if(min > max)
-			return new float[0];
+			return new double[0];
 		
 		double absscale = Math.floor(Math.log10(max-min));
 		double relscale = Math.log10(max-min) - absscale;
-		float ticksize = 0;
+		double ticksize = 0;
 		
 		if( relscale < Math.log10(0.2*steps) )
-		    ticksize = (float) (Math.pow(10,absscale)*0.2);
+		    ticksize = (Math.pow(10,absscale)*0.2);
 		else if( relscale < Math.log10(0.5*steps) )
-		    ticksize = (float) (Math.pow(10,absscale)*0.5);
+		    ticksize = (Math.pow(10,absscale)*0.5);
 		else if( relscale < Math.log10(1*steps) )
-		    ticksize = (float) Math.pow(10,absscale)*1;
+		    ticksize = Math.pow(10,absscale)*1;
 		else
-		    ticksize = (float) Math.pow(10,absscale)*2;
+		    ticksize = Math.pow(10,absscale)*2;
 		
-		int start = (int)Math.ceil(min/ticksize);
-		int stop  = (int)Math.floor(max/ticksize);
+		double start = (int)Math.ceil(min/ticksize);
+		double stop  = (int)Math.floor(max/ticksize);
 
-		float[] ticks = new float[stop-start+1];
+		double[] ticks = new double[(int)(stop-start+1)];
 
-		for(int t=start; t<=stop; t++)
-			ticks[t-start] = (t*ticksize);
+		for(double t=start; t<=stop; t++)
+			ticks[(int)(t-start)] = (t*ticksize);
 		
 		return ticks;
 	}
