+++ src/api/org/jzy3d/plot3d/primitives/axes/layout/renderers/DefaultDecimalTickRenderer.java
@@ -13,7 +13,7 @@ public class DefaultDecimalTickRenderer implements ITickRenderer{
 	}
 	
 	@Override
-	public String format(float value) {
+	public String format(double value) {
 		return Utils.num2str('g', value, precision);
 	}
