+++ src/api/org/jzy3d/plot3d/primitives/axes/layout/renderers/ITickRenderer.java
@@ -1,5 +1,5 @@
 package org.jzy3d.plot3d.primitives.axes.layout.renderers;
 
 public interface ITickRenderer {
-	public String format(float value);
+	public String format(double value);
 }
