+++ src/api/org/jzy3d/plot3d/primitives/axes/layout/renderers/IntegerTickRenderer.java
@@ -5,7 +5,7 @@ public class IntegerTickRenderer implements ITickRenderer{
 	}
 	
 	@Override
-	public String format(float value) {
+	public String format(double value) {
 		return "" + (int)value;
 	}
 }
